feat: Add ring drawing functionality and update main.py to utilize it

+++ run.py
@@ -1,21 +1,29 @@
 import subprocess
 import time
 
-while True:
-    print("Starting main.py...")
-    process = subprocess.Popen(["python", "main.py"])
+try:
+    while True:
+        print("Starting main.py...")
+        process = subprocess.Popen(["python", "main.py"])
 
-    print("Press Enter to kill and restart...")
-    input()
+        print("Press Enter to kill and restart...")
+        op = input("Press 'q' to quit or any other key to restart: ")
+        if op == "q":
+            break
+        print("Killing main.py...")
+        process.terminate()
 
-    print("Killing main.py...")
-    process.terminate()
+        try:
+            process.wait(timeout=5)
+        except subprocess.TimeoutExpired:
+            print("Force killing main.py...")
+            process.kill()
 
-    try:
-        process.wait(timeout=5)
-    except subprocess.TimeoutExpired:
-        print("Force killing main.py...")
-        process.kill()
+        print("Restarting...\n")
+        time.sleep(1)
+except KeyboardInterrupt:
+    process.kill()
+    print("\nExiting...")
 
-    print("Restarting...\n")
-    time.sleep(1)
+process.kill()
+print("Process killed.")
